# backend/pzio/modules/projects/router.py
"""
Responsibilities:
  - Declare routes with correct HTTP methods, paths and status codes.
  - Inject dependencies (DB session, current user) via dependencies.py.
  - Delegate ALL business logic to services.py.
  - Return typed response models with camelCase JSON (response_model_by_alias=True).
"""

from __future__ import annotations

from fastapi import APIRouter, Query, status

from .dependencies import AuthUser, DBSession, ManagerUser
from .schemas import (
    BurndownOut,
    MemberListParams,
    Page,
    ProjectCreate,
    ProjectDetailOut,
    ProjectListParams,
    ProjectMemberCreate,
    ProjectMemberOut,
    ProjectMemberUpdate,
    ProjectOut,
    ProjectUpdate,
    SprintCreate,
    SprintOut,
    SprintUpdate,
)
from . import services

router = APIRouter(tags=["Projects"])

# ---------------------------------------------------------------------------
# PROJECTS
# ---------------------------------------------------------------------------

@router.post(
    "/api/projects",
    response_model=ProjectOut,
    response_model_by_alias=True,
    status_code=status.HTTP_201_CREATED,
    summary="Create a new project",
    description=(
        "Creates a new project owned by the calling user. "
        "The caller is automatically added as a member with the `project_owner` role. "
        "Any authenticated user can create projects."
    ),
)
def create_project(
    payload: ProjectCreate,
    db: DBSession,
    current_user: ManagerUser,
) -> ProjectOut:
    return services.create_project(db, payload, current_user.user_id)


@router.get(
    "/api/projects",
    response_model=Page[ProjectOut],
    response_model_by_alias=True,
    status_code=status.HTTP_200_OK,
    summary="List projects the current user is a member of",
    description=(
        "Returns a paginated list of projects the caller is a member of. "
        "Supports filtering by `status`, free-text `search` over the name, "
        "and sorting via `sortBy` (`created_at`, `name`, `updated_at`) + `sortDirection` (`asc`/`desc`)."
    ),
)
def list_projects(
    db: DBSession,
    current_user: AuthUser,
    status_filter: str | None = Query(default=None, alias="status"),
    search: str | None = Query(default=None),
    sortBy: str = Query(default="created_at"),
    sortDirection: str = Query(default="desc", pattern="^(asc|desc)$"),
    page: int = Query(default=1, ge=1),
    size: int = Query(default=20, ge=1, le=100),
) -> Page[ProjectOut]:
    params = ProjectListParams(
        status=status_filter,
        search=search,
        sortBy=sortBy,
        sortDirection=sortDirection,
        page=page,
        size=size,
    )
    return services.list_projects(db, params, current_user.user_id)


@router.get(
    "/api/projects/{id}",
    response_model=ProjectDetailOut,
    response_model_by_alias=True,
    status_code=status.HTTP_200_OK,
    summary="Get a single project with member and sprint statistics",
    description=(
        "Returns the project together with embedded `stats` (member and sprint counts). "
        "The caller must be a member of the project — otherwise the API returns `404 Not Found` "
        "(membership is not leaked to outsiders)."
    ),
)
def get_project(
    id: int,
    db: DBSession,
    current_user: AuthUser,
) -> ProjectDetailOut:
    return services.get_project(db, id, current_user.user_id)


@router.patch(
    "/api/projects/{id}",
    response_model=ProjectOut,
    response_model_by_alias=True,
    status_code=status.HTTP_200_OK,
    summary="Partially update a project",
    description=(
        "Partial update of a project's name, description or status. "
        "Empty body returns `400`. Requires the caller to be a member of the project."
    ),
)
def update_project(
    id: int,
    payload: ProjectUpdate,
    db: DBSession,
    current_user: AuthUser,
) -> ProjectOut:
    return services.update_project(db, id, payload, current_user.user_id)


@router.delete(
    "/api/projects/{id}",
    status_code=status.HTTP_204_NO_CONTENT,
    summary="Archive (soft-delete) a project",
    description=(
        "Soft-deletes the project by transitioning its `status` to `archived`. "
        "The row stays in the database so historical data (sprints, tasks, comments) remains queryable. "
        "Requires the caller to be a project member."
    ),
)
def delete_project(
    id: int,
    db: DBSession,
    current_user: AuthUser,
) -> None:
    services.delete_project(db, id, current_user.user_id)


# ---------------------------------------------------------------------------
# PROJECT MEMBERS
# ---------------------------------------------------------------------------

@router.post(
    "/api/projects/{id}/members",
    response_model=ProjectMemberOut,
    response_model_by_alias=True,
    status_code=status.HTTP_201_CREATED,
    summary="Add a user to a project (requires project_owner or scrum_master role)",
    description=(
        "Adds an existing user (looked up by email) to the project with one or more project roles. "
        "Returns `404` if the email is unknown, `409` if the user is already a member, "
        "`403` if the caller does not have `project_owner` or `scrum_master`."
    ),
)
def add_member(
    id: int,
    payload: ProjectMemberCreate,
    db: DBSession,
    current_user: AuthUser,
) -> ProjectMemberOut:
    return services.add_member(db, id, payload, current_user.user_id)


@router.get(
    "/api/projects/{id}/members",
    response_model=Page[ProjectMemberOut],
    response_model_by_alias=True,
    status_code=status.HTTP_200_OK,
    summary="List project members with optional role/search filtering",
    description=(
        "Paginated list of project members. Optional `role` filter narrows to a single project role; "
        "optional `search` matches against first name, last name and email (case-insensitive). "
        "Only project members may call this endpoint."
    ),
)
def list_members(
    id: int,
    db: DBSession,
    current_user: AuthUser,
    role: str | None = Query(default=None),
    search: str | None = Query(default=None),
    page: int = Query(default=1, ge=1),
    size: int = Query(default=20, ge=1, le=100),
) -> Page[ProjectMemberOut]:
    params = MemberListParams(role=role, search=search, page=page, size=size)
    return services.list_members(db, id, params, current_user.user_id)


@router.delete(
    "/api/projects/{id}/members/{user_id}",
    status_code=status.HTTP_204_NO_CONTENT,
    summary="Remove a user from a project (requires project_owner or scrum_master role)",
    description=(
        "Removes a member from the project. Requires `project_owner` or `scrum_master`. "
        "Returns `404` if the membership does not exist."
    ),
)
def remove_member(
    id: int,
    user_id: int,
    db: DBSession,
    current_user: AuthUser,
) -> None:
    services.remove_member(db, id, user_id, current_user.user_id)

@router.patch(
    "/api/projects/{id}/members/{user_id}",
    response_model=ProjectMemberOut,
    response_model_by_alias=True,
    status_code=status.HTTP_200_OK,
    summary="Update roles of a project member (requires project_owner or scrum_master role)",
    description=(
        "Replaces the member's project roles with the supplied list (`roles`). "
        "At least one role is required. Returns `403` for callers without `project_owner` or `scrum_master`."
    ),
)
def update_member_roles(
    id: int,
    user_id: int,
    payload: ProjectMemberUpdate,
    db: DBSession,
    current_user: AuthUser,
) -> ProjectMemberOut:
    return services.update_member_roles(db, id, user_id, payload, current_user.user_id)


# ---------------------------------------------------------------------------
# SPRINTS
# ---------------------------------------------------------------------------

@router.post(
    "/api/projects/{id}/sprints",
    response_model=SprintOut,
    response_model_by_alias=True,
    status_code=status.HTTP_201_CREATED,
    summary="Create a sprint in a project",
    description=(
        "Creates a sprint with `status=planned`. `endDate` must be strictly later than `startDate`, "
        "otherwise the API returns `400`. Only project members can create sprints."
    ),
)
def create_sprint(
    id: int,
    payload: SprintCreate,
    db: DBSession,
    current_user: AuthUser,
) -> SprintOut:
    return services.create_sprint(db, id, payload, current_user.user_id)


@router.get(
    "/api/projects/{id}/sprints",
    response_model=list[SprintOut],
    response_model_by_alias=True,
    status_code=status.HTTP_200_OK,
    summary="List all sprints for a project ordered by start date",
    description=(
        "Returns every sprint belonging to the project, ordered by `startDate` ascending. "
        "Non-members get `404`."
    ),
)
def list_sprints(
    id: int,
    db: DBSession,
    current_user: AuthUser,
) -> list[SprintOut]:
    return services.list_sprints(db, id, current_user.user_id)


@router.patch(
    "/api/sprints/{id}",
    response_model=SprintOut,
    response_model_by_alias=True,
    status_code=status.HTTP_200_OK,
    summary="Partially update a sprint",
    description=(
        "Partial update of a sprint (name, dates, goal, status). "
        "Use `status=active` to start a sprint and `status=completed` to close it. "
        "Empty body returns `400`."
    ),
)
def update_sprint(
    id: int,
    payload: SprintUpdate,
    db: DBSession,
    current_user: AuthUser,
) -> SprintOut:
    return services.update_sprint(db, id, payload, current_user.user_id)


@router.delete(
    "/api/sprints/{id}",
    status_code=status.HTTP_204_NO_CONTENT,
    summary="Delete a sprint",
    description=(
        "Hard-deletes a sprint. Tasks previously linked to the sprint have their `sprintId` cleared "
        "rather than being deleted."
    ),
)
def delete_sprint(
    id: int,
    db: DBSession,
    current_user: AuthUser,
) -> None:
    services.delete_sprint(db, id, current_user.user_id)


@router.get(
    "/api/sprints/{id}/burndown",
    response_model=BurndownOut,
    response_model_by_alias=True,
    status_code=status.HTTP_200_OK,
    summary="Get burndown chart data for a sprint",
    description=(
        "Returns one data point per day between `startDate` and `endDate` (inclusive). "
        "Each point reports the sum of story points of tasks not yet in `Done`, "
        "letting the frontend plot a classic burndown chart."
    ),
)
def get_burndown(
    id: int,
    db: DBSession,
    current_user: AuthUser,
) -> BurndownOut:
    return services.get_burndown(db, id, current_user.user_id)