# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: PZIO Task Management System
  description: Backend REST API for the team-project task management system. See documents/4.pdf (SAD)
    for the canonical API contract.
  version: 0.1.0
paths:
...
  /api/projects:
    post:
      tags:
      - Projects
      summary: Create a new project
      description: Creates a new project owned by the calling user. The caller is automatically added
        as a member with the `project_owner` role. Any authenticated user can create projects.
      operationId: create_project_api_projects_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - Projects
      summary: List projects the current user is a member of
      description: Returns a paginated list of projects the caller is a member of. Supports filtering
        by `status`, free-text `search` over the name, and sorting via `sortBy` (`created_at`, `name`,
        `updated_at`) + `sortDirection` (`asc`/`desc`).
      operationId: list_projects_api_projects_get
      security:
      - HTTPBearer: []
      parameters:
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      - name: search
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Search
      - name: sortBy
        in: query
        required: false
        schema:
          type: string
          default: created_at
          title: Sortby
      - name: sortDirection
        in: query
        required: false
        schema:
          type: string
          pattern: ^(asc|desc)$
          default: desc
          title: Sortdirection
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: size
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 20
          title: Size
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_ProjectOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/projects/{id}:
    get:
      tags:
      - Projects
      summary: Get a single project with member and sprint statistics
      description: Returns the project together with embedded `stats` (member and sprint counts). The
        caller must be a member of the project — otherwise the API returns `404 Not Found` (membership
        is not leaked to outsiders).
      operationId: get_project_api_projects__id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          title: Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectDetailOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - Projects
      summary: Partially update a project
      description: Partial update of a project's name, description or status. Empty body returns `400`.
        Requires the caller to be a member of the project.
      operationId: update_project_api_projects__id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          title: Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Projects
      summary: Archive (soft-delete) a project
      description: Soft-deletes the project by transitioning its `status` to `archived`. The row stays
        in the database so historical data (sprints, tasks, comments) remains queryable. Requires the
        caller to be a project member.
      operationId: delete_project_api_projects__id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          title: Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/projects/{id}/members:
    post:
      tags:
      - Projects
      summary: Add a user to a project (requires project_owner or scrum_master role)
      description: Adds an existing user (looked up by email) to the project with one or more project
        roles. Returns `404` if the email is unknown, `409` if the user is already a member, `403` if
        the caller does not have `project_owner` or `scrum_master`.
      operationId: add_member_api_projects__id__members_post
      security:
      - HTTPBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          title: Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectMemberCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectMemberOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - Projects
      summary: List project members with optional role/search filtering
      description: Paginated list of project members. Optional `role` filter narrows to a single project
        role; optional `search` matches against first name, last name and email (case-insensitive). Only
        project members may call this endpoint.
      operationId: list_members_api_projects__id__members_get
      security:
      - HTTPBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          title: Id
      - name: role
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Role
      - name: search
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Search
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: size
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 20
          title: Size
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_ProjectMemberOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/projects/{id}/members/{user_id}:
    delete:
      tags:
      - Projects
      summary: Remove a user from a project (requires project_owner or scrum_master role)
      description: Removes a member from the project. Requires `project_owner` or `scrum_master`. Returns
        `404` if the membership does not exist.
      operationId: remove_member_api_projects__id__members__user_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          title: Id
      - name: user_id
        in: path
        required: true
        schema:
          type: integer
          title: User Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - Projects
      summary: Update roles of a project member (requires project_owner or scrum_master role)
      description: Replaces the member's project roles with the supplied list (`roles`). At least one
        role is required. Returns `403` for callers without `project_owner` or `scrum_master`.
      operationId: update_member_roles_api_projects__id__members__user_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          title: Id
      - name: user_id
        in: path
        required: true
        schema:
          type: integer
          title: User Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectMemberUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectMemberOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/projects/{id}/sprints:
    post:
      tags:
      - Projects
      summary: Create a sprint in a project
      description: Creates a sprint with `status=planned`. `endDate` must be strictly later than `startDate`,
        otherwise the API returns `400`. Only project members can create sprints.
      operationId: create_sprint_api_projects__id__sprints_post
      security:
      - HTTPBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          title: Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SprintCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SprintOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - Projects
      summary: List all sprints for a project ordered by start date
      description: Returns every sprint belonging to the project, ordered by `startDate` ascending. Non-members
        get `404`.
      operationId: list_sprints_api_projects__id__sprints_get
      security:
      - HTTPBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          title: Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SprintOut'
                title: Response List Sprints Api Projects  Id  Sprints Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/sprints/{id}:
    patch:
      tags:
      - Projects
      summary: Partially update a sprint
      description: Partial update of a sprint (name, dates, goal, status). Use `status=active` to start
        a sprint and `status=completed` to close it. Empty body returns `400`.
      operationId: update_sprint_api_sprints__id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          title: Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SprintUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SprintOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Projects
      summary: Delete a sprint
      description: Hard-deletes a sprint. Tasks previously linked to the sprint have their `sprintId`
        cleared rather than being deleted.
      operationId: delete_sprint_api_sprints__id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          title: Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/sprints/{id}/burndown:
    get:
      tags:
      - Projects
      summary: Get burndown chart data for a sprint
      description: Returns one data point per day between `startDate` and `endDate` (inclusive). Each
        point reports the sum of story points of tasks not yet in `Done`, letting the frontend plot a
        classic burndown chart.
      operationId: get_burndown_api_sprints__id__burndown_get
      security:
      - HTTPBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          title: Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BurndownOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    Page_ProjectMemberOut_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/ProjectMemberOut'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        page:
          type: integer
          title: Page
        size:
          type: integer
          title: Size
      type: object
      required:
      - items
      - total
      - page
      - size
      title: Page[ProjectMemberOut]
    Page_ProjectOut_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/ProjectOut'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        page:
          type: integer
          title: Page
        size:
          type: integer
          title: Size
      type: object
      required:
      - items
      - total
      - page
      - size
      title: Page[ProjectOut]
...
    ProjectCreate:
      properties:
        name:
          type: string
          maxLength: 255
          minLength: 1
          title: Name
        description:
          anyOf:
          - type: string
            maxLength: 5000
          - type: 'null'
          title: Description
      type: object
      required:
      - name
      title: ProjectCreate
      description: POST /api/projects - request body.
      example:
        description: Migrate the legacy Apollo billing engine to the new microservice platform.
        name: Apollo Migration
    ProjectDetailOut:
      properties:
        projectId:
          type: integer
          title: Projectid
        name:
          type: string
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        status:
          $ref: '#/components/schemas/ProjectStatus'
        currentUserRoles:
          items:
            $ref: '#/components/schemas/ProjectRole'
          type: array
          title: Currentuserroles
        createdAt:
          type: string
          format: date-time
          title: Createdat
        updatedAt:
          type: string
          format: date-time
          title: Updatedat
        stats:
          $ref: '#/components/schemas/ProjectStats'
      type: object
      required:
      - projectId
      - name
      - status
      - createdAt
      - updatedAt
      - stats
      title: ProjectDetailOut
      description: Extended project response (single GET) with stats.
      example:
        createdAt: '2026-05-14T09:30:00Z'
        currentUserRoles:
        - project_owner
        description: Migrate the legacy Apollo billing engine to the new microservice platform.
        name: Apollo Migration
        projectId: 7
        stats:
          memberCount: 6
          sprintCount: 3
        status: active
        updatedAt: '2026-05-14T09:30:00Z'
    ProjectMemberCreate:
      properties:
        email:
          type: string
          maxLength: 255
          pattern: ^[^@]+@[^@]+\.[^@]+$
          title: Email
        roles:
          items:
            $ref: '#/components/schemas/ProjectRole'
          type: array
          minItems: 1
          title: Roles
      type: object
      required:
      - email
      - roles
      title: ProjectMemberCreate
      description: POST /api/projects/{id}/members – request body.
      example:
        email: bob@example.com
        roles:
        - developer
        - qa_engineer
    ProjectMemberOut:
      properties:
        id:
          type: integer
          title: Id
        projectId:
          type: integer
          title: Projectid
        userId:
          type: integer
          title: Userid
        roles:
          items:
            $ref: '#/components/schemas/ProjectRole'
          type: array
          title: Roles
        joinedAt:
          type: string
          format: date-time
          title: Joinedat
        firstName:
          anyOf:
          - type: string
          - type: 'null'
          title: Firstname
        lastName:
          anyOf:
          - type: string
          - type: 'null'
          title: Lastname
        email:
          type: string
          title: Email
      type: object
      required:
      - id
      - projectId
      - userId
      - roles
      - joinedAt
      - email
      title: ProjectMemberOut
      description: Response body for a single project member.
      example:
        email: bob@example.com
        firstName: Bob
        id: 15
        joinedAt: '2026-05-14T10:00:00Z'
        lastName: Brown
        projectId: 7
        roles:
        - developer
        userId: 88
    ProjectMemberUpdate:
      properties:
        roles:
          items:
            $ref: '#/components/schemas/ProjectRole'
          type: array
          minItems: 1
          title: Roles
      type: object
      required:
      - roles
      title: ProjectMemberUpdate
      description: PATCH /api/projects/{id}/members/ – request body for role update.
      example:
        roles:
        - scrum_master
    ProjectOut:
      properties:
        projectId:
          type: integer
          title: Projectid
        name:
          type: string
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        status:
          $ref: '#/components/schemas/ProjectStatus'
        currentUserRoles:
          items:
            $ref: '#/components/schemas/ProjectRole'
          type: array
          title: Currentuserroles
        createdAt:
          type: string
          format: date-time
          title: Createdat
        updatedAt:
          type: string
          format: date-time
          title: Updatedat
      type: object
      required:
      - projectId
      - name
      - status
      - createdAt
      - updatedAt
      title: ProjectOut
      description: Standard project response body.
      example:
        createdAt: '2026-05-14T09:30:00Z'
        currentUserRoles:
        - project_owner
        description: Migrate the legacy Apollo billing engine to the new microservice platform.
        name: Apollo Migration
        projectId: 7
        status: active
        updatedAt: '2026-05-14T09:30:00Z'
    ProjectRole:
      type: string
      enum:
      - developer
      - qa
      - scrum_master
      - project_owner
      - maintainer
      title: ProjectRole
...
    ProjectStatus:
      type: string
      enum:
      - active
      - archived
      title: ProjectStatus
    ProjectUpdate:
      properties:
        name:
          anyOf:
          - type: string
            maxLength: 255
            minLength: 1
          - type: 'null'
          title: Name
        description:
          anyOf:
          - type: string
            maxLength: 5000
          - type: 'null'
          title: Description
        status:
          anyOf:
          - $ref: '#/components/schemas/ProjectStatus'
          - type: 'null'
      type: object
      title: ProjectUpdate
      description: PATCH /api/projects/{id} - all fields optional (partial update).
      example:
        name: Apollo Migration (Phase 2)
        status: active
    SprintCreate:
      properties:
        name:
          type: string
          maxLength: 255
          minLength: 1
          title: Name
        startDate:
          type: string
          format: date-time
          title: Startdate
        endDate:
          type: string
          format: date-time
          title: Enddate
        goal:
          anyOf:
          - type: string
            maxLength: 1500
          - type: 'null'
          title: Goal
      type: object
      required:
      - name
      - startDate
      - endDate
      title: SprintCreate
      description: POST /api/projects/{id}/sprints - request body.
      example:
        endDate: '2026-05-29T00:00:00Z'
        goal: Ship the auth flow and basic backlog UI.
        name: Sprint 1
        startDate: '2026-05-15T00:00:00Z'
...
    SprintStatus:
      type: string
      enum:
      - planned
      - active
      - completed
      title: SprintStatus
    SprintUpdate:
      properties:
        name:
          anyOf:
          - type: string
            maxLength: 255
            minLength: 1
          - type: 'null'
          title: Name
        startDate:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Startdate
        endDate:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Enddate
        status:
          anyOf:
          - $ref: '#/components/schemas/SprintStatus'
          - type: 'null'
        goal:
          anyOf:
          - type: string
            maxLength: 1500
          - type: 'null'
          title: Goal
      type: object
      title: SprintUpdate
      description: PATCH /api/sprints/{id} - all fields optional.
      example:
        goal: Updated goal mid-sprint.
        status: active
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
...